Moved the logging implementation to log.cpp
